Comments auth at Controller and View

+++ app/Http/Controllers/CommentsController.php
@@ -8,7 +8,12 @@ use App\Comment;
 
 class CommentsController extends Controller
 {
-    public function store(Post $post)
+    public function __construct()
+	{
+		$this->middleware(['auth']);
+	}
+	
+	public function store(Post $post)
     {
         $this->validate(request(), [
             'body' => 'required|min:2',
@@ -21,8 +26,18 @@ class CommentsController extends Controller
     
     public function destroy(Comment $comment)
     {
-        $comment->delete();
-        
+        $user = auth()->user();
+		
+		if($user->isAdmin() || $user->ownes($comment->post) || $user->ownes($comment)) {
+			
+			$comment->delete();
+			
+		} else {
+			
+			flash()->error('Not Allowed', 'You have no permission to delete the comment');
+			
+		}
+		
         return back();
     }
